ats-scanner/requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
click==8.1.8
blinker==1.9.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
python-docx==1.1.2
pdfminer.six==20231228
pdfplumber==0.11.5
groq==0.18.0
fpdf==1.7.2
